# fix terrible bug and inline GC for expiring cache