--- a/staging/src/k8s.io/apiserver/pkg/authentication/token/cache/cache_simple.go
+++ b/staging/src/k8s.io/apiserver/pkg/authentication/token/cache/cache_simple.go
@@ -17,7 +17,6 @@ limitations under the License.
 package cache
 
 import (
-	"context"
 	"time"
 
 	utilcache "k8s.io/apimachinery/pkg/util/cache"
@@ -28,10 +27,8 @@ type simpleCache struct {
 	cache *utilcache.Expiring
 }
 
-func newSimpleCache(ctx context.Context, clock clock.Clock) cache {
-	c := &simpleCache{cache: utilcache.NewExpiringWithClock(clock)}
-	go c.cache.Run(ctx)
-	return c
+func newSimpleCache(clock clock.Clock) cache {
+	return &simpleCache{cache: utilcache.NewExpiringWithClock(clock)}
 }
 
 func (c *simpleCache) get(key string) (*cacheRecord, bool) {