I fudged cleanup of this comment:

https://github.com/kubernetes/kubernetes/pull/84424#discussion_r345346041

It needs to be the opposite of this:

https://github.com/kubernetes/kubernetes/blob/f2469a7a2c75a78b8c0b0514652511d105d2b2e7/staging/src/k8s.io/apimachinery/pkg/util/cache/expiring.go#L75

I've made the expiration test check cache size rather than just calling Get.

This change also inlines GC, removing the background goroutine. This appears to make virtually no difference in performance. I'm happy to separate out the bug fix as well.

Before (including the fix) and after (including fix plus GC inlining) benchmarks for 1s and 10s:

#### -benchtime=1s
Mostly loading.
```
$ benchstat before.1s.log after.1s.log 
name                                                  old time/op     new time/op     delta
CachedTokenAuthenticator/tokens=100_threads=1-60          661ns ± 6%      655ns ±12%     ~     (p=0.797 n=9+10)
CachedTokenAuthenticator/tokens=100_threads=16-60         594ns ±15%      502ns ±17%  -15.43%  (p=0.002 n=10+9)
CachedTokenAuthenticator/tokens=100_threads=256-60        649ns ±35%      686ns ±20%     ~     (p=0.387 n=9+9)
CachedTokenAuthenticator/tokens=500_threads=1-60          376ns ± 2%      387ns ± 2%   +2.98%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=500_threads=16-60         430ns ±17%      408ns ±14%     ~     (p=0.160 n=10+10)
CachedTokenAuthenticator/tokens=500_threads=256-60        843ns ±21%      674ns ±18%  -20.08%  (p=0.002 n=9+9)
CachedTokenAuthenticator/tokens=2500_threads=1-60         364ns ± 1%      361ns ± 1%     ~     (p=0.105 n=10+10)
CachedTokenAuthenticator/tokens=2500_threads=16-60        447ns ±13%      432ns ±13%     ~     (p=0.342 n=10+10)
CachedTokenAuthenticator/tokens=2500_threads=256-60      1.22µs ±25%     1.23µs ±30%     ~     (p=0.853 n=10+10)
CachedTokenAuthenticator/tokens=12500_threads=1-60        446ns ± 1%      407ns ± 1%   -8.82%  (p=0.000 n=8+9)
CachedTokenAuthenticator/tokens=12500_threads=16-60       602ns ± 8%      600ns ± 7%     ~     (p=0.838 n=10+10)
CachedTokenAuthenticator/tokens=12500_threads=256-60     12.1µs ± 8%     12.7µs ± 2%     ~     (p=0.278 n=10+9)
CachedTokenAuthenticator/tokens=62500_threads=1-60       18.0µs ± 0%     17.9µs ± 0%   -0.17%  (p=0.021 n=9+10)
CachedTokenAuthenticator/tokens=62500_threads=16-60      18.8µs ± 0%     18.8µs ± 0%     ~     (p=0.912 n=10+10)
CachedTokenAuthenticator/tokens=62500_threads=256-60     25.0µs ± 1%     25.1µs ± 1%     ~     (p=0.684 n=10+10)

name                                                  old lookups/op  new lookups/op  delta
CachedTokenAuthenticator/tokens=100_threads=1-60           0.00 ±12%       0.00 ±14%     ~     (p=0.754 n=10+10)
CachedTokenAuthenticator/tokens=100_threads=16-60          0.00 ±49%      0.00 ±103%     ~     (p=0.725 n=10+10)
CachedTokenAuthenticator/tokens=100_threads=256-60         0.01 ±46%       0.01 ±50%     ~     (p=0.450 n=9+9)
CachedTokenAuthenticator/tokens=500_threads=1-60           0.00 ± 3%       0.00 ± 5%     ~     (p=0.090 n=9+10)
CachedTokenAuthenticator/tokens=500_threads=16-60          0.00 ±59%       0.00 ±12%  -34.78%  (p=0.004 n=10+10)
CachedTokenAuthenticator/tokens=500_threads=256-60         0.02 ±20%       0.02 ±68%  -30.19%  (p=0.007 n=8+10)
CachedTokenAuthenticator/tokens=2500_threads=1-60          0.00 ± 2%       0.00 ± 3%   -5.10%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=2500_threads=16-60         0.00 ±24%       0.00 ±16%     ~     (p=0.109 n=10+10)
CachedTokenAuthenticator/tokens=2500_threads=256-60        0.03 ±25%       0.04 ±27%     ~     (p=0.853 n=10+10)
CachedTokenAuthenticator/tokens=12500_threads=1-60         0.01 ± 1%       0.01 ± 1%  -15.13%  (p=0.000 n=8+9)
CachedTokenAuthenticator/tokens=12500_threads=16-60        0.01 ± 7%       0.01 ±12%     ~     (p=0.221 n=10+10)
CachedTokenAuthenticator/tokens=12500_threads=256-60       0.40 ± 7%       0.42 ± 1%     ~     (p=0.055 n=10+9)
CachedTokenAuthenticator/tokens=62500_threads=1-60         0.65 ± 0%       0.65 ± 0%   -0.12%  (p=0.029 n=7+10)
CachedTokenAuthenticator/tokens=62500_threads=16-60        0.68 ± 0%       0.68 ± 0%     ~     (p=0.735 n=10+10)
CachedTokenAuthenticator/tokens=62500_threads=256-60       0.88 ± 1%       0.88 ± 1%     ~     (p=0.929 n=10+10)

name                                                  old alloc/op    new alloc/op    delta
CachedTokenAuthenticator/tokens=100_threads=1-60           300B ± 0%       300B ± 0%     ~     (p=1.000 n=10+10)
CachedTokenAuthenticator/tokens=100_threads=16-60          302B ± 0%       302B ± 0%     ~     (all equal)
CachedTokenAuthenticator/tokens=100_threads=256-60         345B ± 5%       349B ± 5%     ~     (p=0.266 n=9+9)
CachedTokenAuthenticator/tokens=500_threads=1-60           297B ± 0%       297B ± 0%     ~     (all equal)
CachedTokenAuthenticator/tokens=500_threads=16-60          299B ± 0%       299B ± 0%     ~     (all equal)
CachedTokenAuthenticator/tokens=500_threads=256-60         354B ± 3%       344B ± 5%   -2.86%  (p=0.006 n=9+9)
CachedTokenAuthenticator/tokens=2500_threads=1-60          297B ± 0%       297B ± 0%     ~     (all equal)
CachedTokenAuthenticator/tokens=2500_threads=16-60         298B ± 0%       299B ± 0%   +0.17%  (p=0.033 n=10+10)
CachedTokenAuthenticator/tokens=2500_threads=256-60        368B ±10%       397B ± 6%   +7.74%  (p=0.006 n=10+10)
CachedTokenAuthenticator/tokens=12500_threads=1-60         298B ± 0%       266B ± 0%  -10.74%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=12500_threads=16-60        302B ± 0%       302B ± 0%     ~     (all equal)
CachedTokenAuthenticator/tokens=12500_threads=256-60     1.39kB ± 8%     1.44kB ± 2%     ~     (p=0.189 n=10+9)
CachedTokenAuthenticator/tokens=62500_threads=1-60         566B ± 0%       566B ± 0%     ~     (p=0.549 n=8+10)
CachedTokenAuthenticator/tokens=62500_threads=16-60        663B ± 0%       663B ± 0%     ~     (p=0.452 n=10+10)
CachedTokenAuthenticator/tokens=62500_threads=256-60     2.66kB ± 9%     2.69kB ± 4%     ~     (p=0.811 n=10+10)

name                                                  old allocs/op   new allocs/op   delta
CachedTokenAuthenticator/tokens=100_threads=1-60           11.0 ± 0%       11.0 ± 0%     ~     (all equal)
CachedTokenAuthenticator/tokens=100_threads=16-60          11.0 ± 0%       11.0 ± 0%     ~     (all equal)
CachedTokenAuthenticator/tokens=100_threads=256-60         11.0 ± 0%       11.0 ± 0%     ~     (all equal)
CachedTokenAuthenticator/tokens=500_threads=1-60           11.0 ± 0%       11.0 ± 0%     ~     (all equal)
CachedTokenAuthenticator/tokens=500_threads=16-60          11.0 ± 0%       11.0 ± 0%     ~     (all equal)
CachedTokenAuthenticator/tokens=500_threads=256-60         11.0 ± 0%       11.0 ± 0%     ~     (all equal)
CachedTokenAuthenticator/tokens=2500_threads=1-60          11.0 ± 0%       11.0 ± 0%     ~     (all equal)
CachedTokenAuthenticator/tokens=2500_threads=16-60         11.0 ± 0%       11.0 ± 0%     ~     (all equal)
CachedTokenAuthenticator/tokens=2500_threads=256-60        10.4 ± 6%       11.4 ± 5%   +9.62%  (p=0.002 n=10+10)
CachedTokenAuthenticator/tokens=12500_threads=1-60         11.0 ± 0%       10.0 ± 0%   -9.09%  (p=0.000 n=10+10)
CachedTokenAuthenticator/tokens=12500_threads=16-60        11.0 ± 0%       11.0 ± 0%     ~     (all equal)
CachedTokenAuthenticator/tokens=12500_threads=256-60       14.0 ± 0%       14.0 ± 0%     ~     (all equal)
CachedTokenAuthenticator/tokens=62500_threads=1-60         15.0 ± 0%       15.0 ± 0%     ~     (all equal)
CachedTokenAuthenticator/tokens=62500_threads=16-60        15.0 ± 0%       15.0 ± 0%     ~     (all equal)
CachedTokenAuthenticator/tokens=62500_threads=256-60       16.0 ± 0%       17.0 ± 0%   +6.25%  (p=0.000 n=10+10)
```

#### -benchtime=10s
Long enough to test caching and GC.
```
$ benchstat before.10s.log after.10s.log 
name                                                  old time/op     new time/op     delta
CachedTokenAuthenticator/tokens=100_threads=1-60          650ns ± 5%      632ns ±10%    ~     (p=0.690 n=5+5)
CachedTokenAuthenticator/tokens=100_threads=16-60         513ns ± 8%      500ns ± 6%    ~     (p=0.516 n=5+5)
CachedTokenAuthenticator/tokens=100_threads=256-60        744ns ±15%      701ns ±14%    ~     (p=0.548 n=5+5)
CachedTokenAuthenticator/tokens=500_threads=1-60          374ns ± 1%      373ns ± 0%    ~     (p=0.968 n=5+5)
CachedTokenAuthenticator/tokens=500_threads=16-60         428ns ± 5%      421ns ± 4%    ~     (p=0.421 n=5+5)
CachedTokenAuthenticator/tokens=500_threads=256-60        814ns ±11%      855ns ±11%    ~     (p=0.460 n=5+5)
CachedTokenAuthenticator/tokens=2500_threads=1-60         346ns ± 1%      352ns ± 1%  +1.79%  (p=0.016 n=5+5)
CachedTokenAuthenticator/tokens=2500_threads=16-60        418ns ± 7%      415ns ±16%    ~     (p=0.278 n=5+5)
CachedTokenAuthenticator/tokens=2500_threads=256-60       980ns ±10%      946ns ± 7%    ~     (p=0.548 n=5+5)
CachedTokenAuthenticator/tokens=12500_threads=1-60        309ns ± 1%      309ns ± 1%    ~     (p=0.913 n=5+5)
CachedTokenAuthenticator/tokens=12500_threads=16-60       494ns ± 9%      509ns ± 7%    ~     (p=0.516 n=5+5)
CachedTokenAuthenticator/tokens=12500_threads=256-60     1.60µs ± 2%     1.57µs ± 1%  -1.89%  (p=0.040 n=5+5)
CachedTokenAuthenticator/tokens=62500_threads=1-60        609ns ± 0%      616ns ± 0%  +1.12%  (p=0.008 n=5+5)
CachedTokenAuthenticator/tokens=62500_threads=16-60       710ns ± 0%      742ns ± 1%  +4.54%  (p=0.029 n=4+4)
CachedTokenAuthenticator/tokens=62500_threads=256-60     4.52µs ± 2%     4.51µs ± 1%    ~     (p=1.000 n=5+5)

name                                                  old lookups/op  new lookups/op  delta
CachedTokenAuthenticator/tokens=100_threads=1-60           0.00 ±10%       0.00 ± 7%    ~     (p=0.310 n=5+5)
CachedTokenAuthenticator/tokens=100_threads=16-60          0.00 ±29%       0.00 ±28%    ~     (p=0.421 n=5+5)
CachedTokenAuthenticator/tokens=100_threads=256-60         0.02 ±24%       0.02 ±24%    ~     (p=0.548 n=5+5)
CachedTokenAuthenticator/tokens=500_threads=1-60           0.00 ± 2%       0.00 ± 6%    ~     (p=0.690 n=5+5)
CachedTokenAuthenticator/tokens=500_threads=16-60          0.00 ± 9%       0.00 ± 7%    ~     (p=0.151 n=5+5)
CachedTokenAuthenticator/tokens=500_threads=256-60         0.02 ±17%       0.03 ±14%    ~     (p=0.310 n=5+5)
CachedTokenAuthenticator/tokens=2500_threads=1-60          0.00 ± 1%       0.00 ± 1%  +1.78%  (p=0.008 n=5+5)
CachedTokenAuthenticator/tokens=2500_threads=16-60         0.00 ±21%       0.00 ±23%    ~     (p=0.690 n=5+5)
CachedTokenAuthenticator/tokens=2500_threads=256-60        0.03 ±12%       0.03 ± 7%    ~     (p=0.595 n=5+5)
CachedTokenAuthenticator/tokens=12500_threads=1-60         0.00 ± 1%       0.00 ± 1%    ~     (p=0.841 n=5+5)
CachedTokenAuthenticator/tokens=12500_threads=16-60        0.01 ±16%       0.01 ±12%    ~     (p=0.500 n=5+5)
CachedTokenAuthenticator/tokens=12500_threads=256-60       0.05 ± 2%       0.05 ± 1%    ~     (p=0.151 n=5+5)
CachedTokenAuthenticator/tokens=62500_threads=1-60         0.02 ± 0%       0.02 ± 0%  +0.92%  (p=0.016 n=4+5)
CachedTokenAuthenticator/tokens=62500_threads=16-60        0.02 ± 1%       0.02 ±11%    ~     (p=0.746 n=5+5)
CachedTokenAuthenticator/tokens=62500_threads=256-60       0.16 ± 2%       0.16 ± 1%    ~     (p=0.873 n=5+5)

name                                                  old alloc/op    new alloc/op    delta
CachedTokenAuthenticator/tokens=100_threads=1-60           299B ± 0%       299B ± 0%    ~     (all equal)
CachedTokenAuthenticator/tokens=100_threads=16-60          293B ± 0%       293B ± 0%    ~     (all equal)
CachedTokenAuthenticator/tokens=100_threads=256-60         303B ± 0%       304B ± 0%    ~     (p=0.556 n=5+4)
CachedTokenAuthenticator/tokens=500_threads=1-60           296B ± 0%       296B ± 0%    ~     (all equal)
CachedTokenAuthenticator/tokens=500_threads=16-60          296B ± 0%       296B ± 0%    ~     (all equal)
CachedTokenAuthenticator/tokens=500_threads=256-60         305B ± 0%       305B ± 0%    ~     (p=0.444 n=5+5)
CachedTokenAuthenticator/tokens=2500_threads=1-60          297B ± 0%       297B ± 0%    ~     (all equal)
CachedTokenAuthenticator/tokens=2500_threads=16-60         297B ± 0%       297B ± 0%    ~     (all equal)
CachedTokenAuthenticator/tokens=2500_threads=256-60        307B ± 0%       306B ± 0%    ~     (p=0.095 n=4+5)
CachedTokenAuthenticator/tokens=12500_threads=1-60         264B ± 0%       264B ± 0%    ~     (all equal)
CachedTokenAuthenticator/tokens=12500_threads=16-60        297B ± 0%       297B ± 0%    ~     (all equal)
CachedTokenAuthenticator/tokens=12500_threads=256-60       315B ± 0%       314B ± 0%    ~     (p=0.095 n=5+4)
CachedTokenAuthenticator/tokens=62500_threads=1-60         299B ± 0%       299B ± 0%    ~     (all equal)
CachedTokenAuthenticator/tokens=62500_threads=16-60        300B ± 0%       300B ± 0%    ~     (all equal)
CachedTokenAuthenticator/tokens=62500_threads=256-60       349B ± 0%       349B ± 0%    ~     (p=0.444 n=5+5)

name                                                  old allocs/op   new allocs/op   delta
CachedTokenAuthenticator/tokens=100_threads=1-60           11.0 ± 0%       11.0 ± 0%    ~     (all equal)
CachedTokenAuthenticator/tokens=100_threads=16-60          11.0 ± 0%       11.0 ± 0%    ~     (all equal)
CachedTokenAuthenticator/tokens=100_threads=256-60         11.0 ± 0%       11.0 ± 0%    ~     (all equal)
CachedTokenAuthenticator/tokens=500_threads=1-60           11.0 ± 0%       11.0 ± 0%    ~     (all equal)
CachedTokenAuthenticator/tokens=500_threads=16-60          11.0 ± 0%       11.0 ± 0%    ~     (all equal)
CachedTokenAuthenticator/tokens=500_threads=256-60         11.0 ± 0%       11.0 ± 0%    ~     (all equal)
CachedTokenAuthenticator/tokens=2500_threads=1-60          11.0 ± 0%       11.0 ± 0%    ~     (all equal)
CachedTokenAuthenticator/tokens=2500_threads=16-60         11.0 ± 0%       11.0 ± 0%    ~     (all equal)
CachedTokenAuthenticator/tokens=2500_threads=256-60        11.0 ± 0%       11.0 ± 0%    ~     (all equal)
CachedTokenAuthenticator/tokens=12500_threads=1-60         10.0 ± 0%       10.0 ± 0%    ~     (all equal)
CachedTokenAuthenticator/tokens=12500_threads=16-60        11.0 ± 0%       11.0 ± 0%    ~     (all equal)
CachedTokenAuthenticator/tokens=12500_threads=256-60       12.0 ± 0%       12.0 ± 0%    ~     (all equal)
CachedTokenAuthenticator/tokens=62500_threads=1-60         11.0 ± 0%       11.0 ± 0%    ~     (all equal)
CachedTokenAuthenticator/tokens=62500_threads=16-60        11.0 ± 0%       11.0 ± 0%    ~     (all equal)
CachedTokenAuthenticator/tokens=62500_threads=256-60       12.0 ± 0%       12.0 ± 0%    ~     (all equal)
```

/kind bug
/sig auth
Fixes: #85317

```release-note
NONE
```