@lavalamp 2019-11-15
huh, I'm surprised the goroutine removal did so little.