@lavalamp 2019-11-16
Can you add a test that verifies that a call to gc removes things it should and doesn't remove things it shouldn't?