@mikedanese 2019-11-16
> huh, I'm surprised the goroutine removal did so little.

I've noticed that GC is actually very cheap vs the time spent in Set() (even before the bug was introduced). I'm guessing that the minimal time spent in GC spreads pretty evenly when inlined. Will validate.

> Can you add a test that verifies that a call to gc removes things it should and doesn't remove things it shouldn't?

Yup, will do.

> Also how do you generate the awesome stats from the benchmarks?

Benchmarks run on a c2-standard-60:

```
[master]$ go test -c -o old ./vendor/k8s.io/apiserver/pkg/authentication/token/cache
[expcache]$ go test -c -o new ./vendor/k8s.io/apiserver/pkg/authentication/token/cache
```

```
#!/bin/bash

set -o errexit
set -o xtrace
set -o nounset
set -o pipefail

SHARED="-test.v -test.benchmem -test.bench='CachedTokenAuthenticator/.*' -test.run='^$' -test.timeout=1h"
SHORT="${SHARED} -test.count=10 -test.benchtime=1s"
LONG="${SHARED} -test.count=5 -test.benchtime=10s"

gcloud compute scp --zone us-central1-c old new perf:~/

gcloud compute ssh  perf --zone us-central1-c \
  --command="./old ${SHORT}" \
  | tee before.1s.log

gcloud compute ssh  perf --zone us-central1-c \
  --command="./new ${SHORT}" \
  | tee after.1s.log

bencstat before.1s.log after.1s.log | tee 1s.diff.log

gcloud compute ssh  perf --zone us-central1-c \
  --command="./old ${LONG}" \
  | tee before.10s.log

gcloud compute ssh  perf --zone us-central1-c \
  --command="./new ${LONG}" \
  | tee after.10s.log

bencstat before.10s.log after.10s.log | tee 1s.diff.log
```

https://godoc.org/golang.org/x/perf/cmd/benchstat