@lavalamp 2019-11-16
> I've noticed that GC is actually very cheap vs the time spent in Set() (even before the bug was introduced). I'm guessing that the minimal time spent in GC spreads pretty evenly when inlined. Will validate.

Yeah, I was hoping for an improvement, but even though it's a no-op removing the goroutine is still a large code complexity improvement, I think.

Do you want to add a test to this PR or in a followup?

Thanks for the stats write-up, I'll be referring to that in the future.