@mikedanese 2019-11-16
> Do you want to add a test to this PR or in a followup?

No, I can do it now.